ItemData refactored + added tests and bench
